add download file when file is corrupted
