+++ src/lib/controllers/texts-controller.js
@@ -226,18 +226,9 @@ export default class TextsController {
       })
       return
     }
-    let alignment
 
-    try {
-      alignment = await UploadController.upload('indexedDBUpload', alData)
-      return alignment
-    } catch (error) {
-      console.error(L10nSingleton.getMsgS('TEXTS_CONTROLLER_INCORRECT_DB_DATA'))
-      NotificationSingleton.addNotification({
-        text: L10nSingleton.getMsgS('TEXTS_CONTROLLER_INCORRECT_DB_DATA'),
-        type: NotificationSingleton.types.ERROR
-      })
-    }
+    const alignment = await UploadController.upload('indexedDBUpload', alData)
+    return alignment
   }
 
   uploadDataFromFile (fileData, tokenizerOptionValue, extension) {
